Split provider and consumer mdib tests
