--- tests/test_mdib.py
@@ -1,259 +1,0 @@
-"""Tests for mdib handling."""
-
-import unittest
-import unittest.mock
-from dataclasses import dataclass
-from pathlib import Path
-
-from lxml import etree
-
-from sdc11073 import definitions_sdc
-from sdc11073.exceptions import ApiUsageError
-from sdc11073.mdib import ConsumerMdib, ProviderMdib
-from sdc11073.xml_types import pm_qnames as pm
-
-MDIB_FOLDER = Path(__file__).parent
-MDIB_TWO_MDS_PATH = MDIB_FOLDER / 'mdib_two_mds.xml'
-MDIB_TNS_PATH = MDIB_FOLDER / 'mdib_tns.xml'
-
-
-class TestMdib(unittest.TestCase):
-    def test_mdib_tns(self):
-        # verify that a mdib with participant model as default namespace can be handled.
-        # if creation does not raise any exception, all should be fine.
-        device_mdib_container = ProviderMdib.from_mdib_file(
-            MDIB_TNS_PATH,
-            protocol_definition=definitions_sdc.SdcV1Definitions,
-        )
-        self.assertTrue(device_mdib_container is not None)
-
-
-class TestMdibTransaction(unittest.TestCase):
-    def setUp(self):
-        self.mdib = ProviderMdib.from_mdib_file(MDIB_TWO_MDS_PATH, protocol_definition=definitions_sdc.SdcV1Definitions)
-
-    def test_create_delete_descriptor(self):
-        with self.mdib.descriptor_transaction() as mgr:
-            parent_descriptor = self.mdib.descriptions.handle.get_one('ch0.vmd0')
-            descriptor_container = self.mdib.data_model.mk_descriptor_container(
-                pm.NumericMetricDescriptor,
-                handle='testHandle',
-                parent_descriptor=parent_descriptor,
-            )
-            state = self.mdib.data_model.mk_state_container(descriptor_container)
-            mgr.add_descriptor(descriptor_container, state_container=state)
-        tr = self.mdib._transaction
-        self.assertEqual(1, len(tr.descr_created))
-        self.assertEqual(1, len(tr.descr_updated))
-        self.assertEqual(2, len(tr.all_states()))
-        descr = self.mdib.descriptions.handle.get_one('testHandle', allow_none=True)
-        state = self.mdib.states.descriptor_handle.get_one('testHandle', allow_none=True)
-        self.assertIsNotNone(descr)
-        self.assertIsNotNone(state)
-
-        with self.mdib.descriptor_transaction() as mgr:
-            mgr.remove_descriptor('testHandle')
-        tr = self.mdib._transaction
-        self.assertEqual(0, len(tr.descr_created))
-        self.assertEqual(1, len(tr.descr_updated))
-        self.assertEqual(1, len(tr.all_states()))
-
-        descr = self.mdib.descriptions.handle.get_one('testHandle', allow_none=True)
-        state = self.mdib.states.descriptor_handle.get_one('testHandle', allow_none=True)
-        self.assertIsNone(descr)
-        self.assertIsNone(state)
-
-    def test_create_descriptor_without_state(self):
-        with self.mdib.descriptor_transaction() as mgr:  # now without state
-            parent_descriptor = self.mdib.descriptions.handle.get_one('ch0.vmd0')
-            descriptor_container = self.mdib.data_model.mk_descriptor_container(
-                pm.NumericMetricDescriptor,
-                handle='testHandle',
-                parent_descriptor=parent_descriptor,
-            )
-            mgr.add_descriptor(descriptor_container)
-        tr = self.mdib._transaction
-        self.assertEqual(1, len(tr.descr_created))
-        self.assertEqual(1, len(tr.descr_updated))
-        self.assertEqual(1, len(tr.all_states()))
-
-    def test_update_descriptor_get_twice(self):
-        with self.mdib.descriptor_transaction() as mgr:
-            metric_descriptor = mgr.get_descriptor('numeric.ch0.vmd0')
-            metric_descriptor.DeterminationPeriod = 29.0
-            state = mgr.get_state('numeric.ch0.vmd0')
-            self.assertEqual(state.DescriptorHandle, 'numeric.ch0.vmd0')
-            self.assertRaises(ValueError, mgr.get_state, 'numeric.ch0.vmd0')  # second get_state call
-            # next call failed due to a previous error
-            self.assertRaises(ApiUsageError, mgr.get_state, 'numeric.ch1.vmd0')
-
-    def test_update_descriptor_wrong_state(self):
-        with self.mdib.descriptor_transaction() as mgr:
-            metric_descriptor = mgr.get_descriptor('numeric.ch0.vmd0')
-            metric_descriptor.DeterminationPeriod = 29.0
-            self.assertRaises(ApiUsageError, mgr.get_state, 'numeric.ch1.vmd0')
-
-    def test_get_mixed_states(self):
-        with self.mdib.metric_state_transaction() as mgr:
-            state = mgr.get_state('numeric.ch0.vmd0')
-            self.assertEqual(state.DescriptorHandle, 'numeric.ch0.vmd0')
-            self.assertRaises(ApiUsageError, mgr.get_state, 'ch0.vmd0')
-
-    def test_activate_operation_argument(self):
-        """Test that pm:ActivateOperationDescriptor/pm:argument/pm:Arg is handled correctly.
-
-        QName as node text is beyond what xml libraries can handle automatically,
-        it must be handled specifically in sdc11073 code.
-        """
-
-        @dataclass
-        class TestData:
-            mdib_text: str
-            expected_qname: etree.QName
-
-        mdib_dummy = """
-        <msg:GetMdibResponse
-        xmlns:msg="http://standards.ieee.org/downloads/11073/11073-10207-2017/message"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         {0}
-         xmlns:pm="http://standards.ieee.org/downloads/11073/11073-10207-2017/participant"
-         xmlns:xsd="http://www.w3.org/2001/XMLSchema"
-         xmlns:ext="http://standards.ieee.org/downloads/11073/11073-10207-2017/extension"
-         MdibVersion="174" SequenceId="urn:uuid:6f4ff7de-6809-4883-9938-54dd4f6f9173">
-        <msg:Mdib MdibVersion="174" SequenceId="urn:uuid:6f4ff7de-6809-4883-9938-54dd4f6f9173">
-            <pm:MdDescription DescriptionVersion="8">
-                <pm:Mds Handle="mds0" DescriptorVersion="5" SafetyClassification="MedA">
-                    <pm:Sco Handle="Sco.mds0" DescriptorVersion="2">
-                        <pm:Operation {1} xsi:type="pm:ActivateOperationDescriptor" Handle="SVO.38.3569"
-                        DescriptorVersion="2" SafetyClassification="MedA" OperationTarget="3569"
-                        MaxTimeToFinish="PT00H00M02S" Retriggerable="false">
-                            <pm:Type Code="193821">
-            <pm:ConceptDescription Lang="en-US">An operation to cancel global all audio pause</pm:ConceptDescription>
-                            </pm:Type>
-                            <pm:Argument>
-                                <pm:ArgName Code="codeForArgumentName"></pm:ArgName>
-                                <pm:Arg {2}>{3}duration</pm:Arg>
-                            </pm:Argument>
-                        </pm:Operation>
-                    </pm:Sco>
-                    <pm:SystemContext Handle="SC.mds0" DescriptorVersion="2">
-                        <ext:Extension>
-                        </ext:Extension>
-                        <pm:PatientContext Handle="PC.mds0" DescriptorVersion="2">
-                        </pm:PatientContext>
-                        <pm:LocationContext Handle="LC.mds0" DescriptorVersion="2">
-                        </pm:LocationContext>
-                    </pm:SystemContext>
-                </pm:Mds>
-            </pm:MdDescription>
-        </msg:Mdib>
-    </msg:GetMdibResponse>"""
-
-        my_prefix = 'my'
-        xsd_prefix = 'xsd'
-        delaration = 'xmlns:{0}="http://www.w3.org/2001/XMLSchema"'
-        delaration_any_uri = 'xmlns:{0}="urn:oid:1.23.3.123.2"'
-        expected_qname_xsd = etree.QName('http://www.w3.org/2001/XMLSchema', 'duration')
-        expected_qname_any_uri = etree.QName('urn:oid:1.23.3.123.2', 'duration')
-
-        mdibs = [
-            TestData(
-                mdib_text=mdib_dummy.format('', '', delaration.format(my_prefix), f'{my_prefix}:'),
-                expected_qname=expected_qname_xsd,
-            ),
-            TestData(
-                mdib_text=mdib_dummy.format('', delaration.format(my_prefix), '', f'{my_prefix}:'),
-                expected_qname=expected_qname_xsd,
-            ),
-            TestData(
-                mdib_text=mdib_dummy.format(delaration.format(my_prefix), '', '', f'{my_prefix}:'),
-                expected_qname=expected_qname_xsd,
-            ),
-            TestData(
-                mdib_text=mdib_dummy.format('', '', delaration.format(xsd_prefix), f'{xsd_prefix}:'),
-                expected_qname=expected_qname_xsd,
-            ),
-            TestData(
-                mdib_text=mdib_dummy.format('', '', 'xmlns="http://www.w3.org/2001/XMLSchema"', ''),
-                expected_qname=expected_qname_xsd,
-            ),
-            TestData(
-                mdib_text=mdib_dummy.format('', '', delaration_any_uri.format(xsd_prefix), f'{xsd_prefix}:'),
-                expected_qname=expected_qname_any_uri,
-            ),
-            TestData(
-                mdib_text=mdib_dummy.format('', delaration_any_uri.format(xsd_prefix), '', f'{xsd_prefix}:'),
-                expected_qname=expected_qname_any_uri,
-            ),
-        ]
-
-        for test_data in mdibs:
-            # parse mdib data into container and reconstruct mdib data back to a msg:GetMdibResponse
-            # so that it can be validated by xml schema validator
-            mdib_text = test_data.mdib_text.encode('utf-8')
-            mdib_container = ProviderMdib.from_string(mdib_text)
-            mdib_node, _ = mdib_container.reconstruct_mdib_with_context_states()
-            arg_nodes = mdib_node.xpath(
-                '//*/pm:Arg',
-                namespaces={'pm': 'http://standards.ieee.org/downloads/11073/11073-10207-2017/participant'},
-            )
-            arg_node = arg_nodes[0]
-            prefix = arg_node.text.split(':')[0]
-            self.assertTrue(prefix in arg_node.nsmap)
-            self.assertEqual(test_data.expected_qname.namespace, arg_node.nsmap[prefix])
-            self.assertEqual(test_data.expected_qname.localname, arg_node.text.split(':')[1])
-
-
-class TestHasNewStateUsableStateVersion(unittest.TestCase):
-    REPORT_NAME = 'SomeReport'
-
-    def setUp(self) -> None:
-        sdc_client = unittest.mock.MagicMock()
-        sdc_client.sdc_definitions = definitions_sdc.SdcV1Definitions
-        sdc_client.log_prefix = ''
-        self.mdib = ConsumerMdib(sdc_client)
-        self.mdib._logger = unittest.mock.MagicMock()
-
-    @staticmethod
-    def _mk_state(state_version: int) -> unittest.mock.MagicMock:
-        state = unittest.mock.MagicMock()
-        state.StateVersion = state_version
-        state.DescriptorHandle = 'my_handle'
-        return state
-
-    def _has_new_state_usable_state_version(self, old_version: int, new_version: int) -> bool:
-        return self.mdib._has_new_state_usable_state_version(
-            self._mk_state(old_version),
-            self._mk_state(new_version),
-            self.REPORT_NAME,
-        )
-
-    def test_incremented_state_version_returns_true(self):
-        self.assertTrue(self._has_new_state_usable_state_version(41, 42))
-
-    def test_missed_state_versions_raises(self):
-        with self.assertRaises(ValueError) as ctx:
-            self._has_new_state_usable_state_version(41, 44)
-        msg = str(ctx.exception)
-        self.assertIn(self.REPORT_NAME, msg)
-        self.assertIn('missed 2 states', msg)
-        self.assertIn('my_handle', msg)
-        self.assertIn('(41->44)', msg)
-
-    def test_decremented_state_version_raises(self):
-        with self.assertRaises(ValueError) as ctx:
-            self._has_new_state_usable_state_version(42, 41)
-        msg = str(ctx.exception)
-        self.assertIn(self.REPORT_NAME, msg)
-        self.assertIn('received older state', msg)
-        self.assertIn('my_handle', msg)
-        self.assertIn('(42->41)', msg)
-
-    def test_repeated_state_version_raises(self):
-        with self.assertRaises(ValueError) as ctx:
-            self._has_new_state_usable_state_version(42, 42)
-        msg = str(ctx.exception)
-        self.assertIn(self.REPORT_NAME, msg)
-        self.assertIn('received older state', msg)
-        self.assertIn('my_handle', msg)
-        self.assertIn('(42->42)', msg)
